set background color of floating divs
If a floating div is next to (and in front of) a div that takes up the
whole screen width, and has a background color set, that color can bleed
through into the floating div. One exampe of this is inline's blogpost
form. Avoid it by setting the background color of the floating element.