my try at improving comment styling
Makes it look more like a blog, but not enough to be confusing, and with
nothing as large as in a blog. Removal of the vertical line under the
subject imho makes it easier to scan through comments as each box is a new
one. Bolding the subject seems to make it stand out enough, especially as
its a link now. (Also considered increasing its font size to 110%.)